# config/openapi.yml
openapi: 3.1.0
info:
  title: Umoja Afya Enterprise EHR
  description: Umoja Afya Tanzania enterprise electronic health record. Includes national
    MPI and patient registration, scheduling/referrals, ADT and unit management, longitudinal
    documentation, orders/results, flowsheets/eMAR, pharmacy, revenue cycle, supply
    chain, quality, public health, remote care, FHIR and audit. Patient-facing portal
    delivery is intentionally deferred from this release.
  version: 11.0.0
paths:
...
  /api/v1/orders:
    get:
      tags:
      - Orders and Results
      summary: List Orders
      operationId: list_orders_api_v1_orders_get
      parameters:
      - name: patient_mpi_id
        in: query
        required: true
        schema:
          type: string
          minLength: 3
          title: Patient Mpi Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Orders and Results
      summary: Create Order
      operationId: create_order_api_v1_orders_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderIn'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/orders/{order_id}/actions:
    post:
      tags:
      - Orders and Results
      summary: Change Order Course
      operationId: change_order_course_api_v1_orders__order_id__actions_post
      parameters:
      - name: order_id
        in: path
        required: true
        schema:
          type: string
          title: Order Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderActionIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/results:
    get:
      tags:
      - Orders and Results
      summary: List Results
      operationId: list_results_api_v1_results_get
      parameters:
      - name: patient_mpi_id
        in: query
        required: true
        schema:
          type: string
          minLength: 3
          title: Patient Mpi Id
      - name: critical_only
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Critical Only
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/results/{result_id}/acknowledge:
    post:
      tags:
      - Orders and Results
      summary: Acknowledge Result
      operationId: acknowledge_result_api_v1_results__result_id__acknowledge_post
      parameters:
      - name: result_id
        in: path
        required: true
        schema:
          type: string
          title: Result Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcknowledgeIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AcknowledgeIn:
      properties:
        actor:
          type: string
          title: Actor
          default: demo-provider
        action_taken:
          anyOf:
          - type: string
          - type: 'null'
          title: Action Taken
      type: object
      title: AcknowledgeIn
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OrderActionIn:
      properties:
        action:
          type: string
          enum:
          - HOLD
          - RESUME
          - CANCEL
          - REINSTATE
          title: Action
        actor:
          type: string
          maxLength: 160
          minLength: 2
          title: Actor
        reason:
          type: string
          maxLength: 1000
          minLength: 3
          title: Reason
      type: object
      required:
      - action
      - actor
      - reason
      title: OrderActionIn
    OrderIn:
      properties:
        encounter_id:
          type: string
          title: Encounter Id
        orderable_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Orderable Code
        order_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Order Type
        order_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Order Name
        priority:
          anyOf:
          - type: string
          - type: 'null'
          title: Priority
        indication:
          anyOf:
          - type: string
          - type: 'null'
          title: Indication
        instructions:
          anyOf:
          - type: string
          - type: 'null'
          title: Instructions
        ordered_by:
          type: string
          title: Ordered By
          default: demo-provider
      type: object
      required:
      - encounter_id
      title: OrderIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...